Fixed a bug which caused problems with passing arguments to R script.

+++ SpeedtestAndExportAndPlotData.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
 
-##	Arguments:
+##	Compulsory Arguments:
 ##	$1: if true, a speedtest will be performed
 ##	$2: if true, data will be exported to csv
 ##	$3: if true, data will be plotted to a png file
+##
+##	Optional Arguments:
 ##	$4: arg1 of R script
 ##	$5: arg2 of R script
 
@@ -19,7 +21,15 @@ fi
 
 if [ "$3" = "1" ] || [ "$3" = "true" ] || [ "$3" = "True" ] || [ "$3" = "TRUE" ]; then
   echo "" # Just because this new line feels good.
-	Rscript "Plotting Data with R.R" "$4" "$5"
+
+  if [ "$#" -eq 3 ]; then
+		Rscript "Plotting Data with R.R"
+	elif [ "$#" -eq 4 ]; then
+	  Rscript "Plotting Data with R.R" $4
+	else
+		Rscript "Plotting Data with R.R" $4 $5
+	fi
+
 	echo "Plotted Data in file."
 	rm Rplots.pdf # Remove the unnecessary pdf file
 fi
